refactor(thales-kms): extract common token service authentication logic
Extracts duplicated HTTP request/response handling from
ClientCertificateTokenService and UserAuthenticationTokenService into
AbstractTokenService base class. Both services share identical patterns
for JSON serialization, HTTP POST execution, response parsing, and error
handling.

Changes:
- Add AbstractTokenService with shared authentication infrastructure
- ClientCertificateTokenService extends AbstractTokenService (78 lines removed)
- UserAuthenticationTokenService extends AbstractTokenService (81 lines removed)
- UserAuthenticationTokenService overrides createBearerToken() to capture refresh tokens

All existing unit and integration tests pass without modification.
