+++ kroxylicious-kms-providers/kroxylicious-kms-provider-thales-ciphertrust-manager/src/main/java/io/kroxylicious/kms/provider/thales/ciphertrust/auth/UserAuthenticationTokenService.java
@@ -6,29 +6,17 @@
 
 package io.kroxylicious.kms.provider.thales.ciphertrust.auth;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
 import java.net.URI;
 import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Objects;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.UnaryOperator;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import io.kroxylicious.kms.provider.thales.ciphertrust.model.AuthRequest;
 import io.kroxylicious.kms.provider.thales.ciphertrust.model.AuthResponse;
-import io.kroxylicious.kms.service.KmsException;
 
 /**
  * Bearer token service for CipherTrust Manager user authentication.
@@ -38,15 +26,10 @@ import io.kroxylicious.kms.service.KmsException;
  * {@link CachingBearerTokenService} for automatic token caching and refresh.
  * </p>
  */
-public class UserAuthenticationTokenService implements BearerTokenService {
+public class UserAuthenticationTokenService extends AbstractTokenService {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(UserAuthenticationTokenService.class);
-    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
-
-    private final URI authEndpoint;
     private final String username;
     private final String password;
-    private final HttpClient client;
     private final AtomicReference<String> refreshToken = new AtomicReference<>();
 
     /**
@@ -63,23 +46,13 @@ public class UserAuthenticationTokenService implements BearerTokenService {
                                           String password,
                                           Duration timeout,
                                           UnaryOperator<HttpClient.Builder> tlsConfigurator) {
-        Objects.requireNonNull(endpointUrl, "endpointUrl cannot be null");
+        super(endpointUrl, timeout, tlsConfigurator);
+
         Objects.requireNonNull(username, "username cannot be null");
         Objects.requireNonNull(password, "password cannot be null");
-        Objects.requireNonNull(timeout, "timeout cannot be null");
-        Objects.requireNonNull(tlsConfigurator, "tlsConfigurator cannot be null");
 
-        this.authEndpoint = endpointUrl.resolve("/api/v1/auth/tokens/");
         this.username = username;
         this.password = password;
-        this.client = createClient(timeout, tlsConfigurator);
-    }
-
-    private HttpClient createClient(Duration timeout, UnaryOperator<HttpClient.Builder> tlsConfigurator) {
-        return tlsConfigurator.apply(HttpClient.newBuilder())
-                .followRedirects(HttpClient.Redirect.NORMAL)
-                .connectTimeout(timeout)
-                .build();
     }
 
     @Override
@@ -113,81 +86,17 @@ public class UserAuthenticationTokenService implements BearerTokenService {
         return authenticate(request, "token refresh");
     }
 
-    private CompletionStage<BearerToken> authenticate(AuthRequest authRequest, String operationType) {
-        try {
-            String requestBody = OBJECT_MAPPER.writeValueAsString(authRequest);
-
-            HttpRequest httpRequest = HttpRequest.newBuilder()
-                    .uri(authEndpoint)
-                    .header("Content-Type", "application/json")
-                    .header("Accept", "application/json")
-                    .POST(HttpRequest.BodyPublishers.ofString(requestBody, StandardCharsets.UTF_8))
-                    .build();
-
-            return client.sendAsync(httpRequest, HttpResponse.BodyHandlers.ofByteArray())
-                    .thenApply(response -> {
-                        if (response.statusCode() == 200) {
-                            return response;
-                        }
-                        else {
-                            String body = new String(response.body(), StandardCharsets.UTF_8);
-                            var logBuilder = LOGGER.atWarn()
-                                    .addKeyValue("operation", operationType)
-                                    .addKeyValue("statusCode", response.statusCode());
-                            if (LOGGER.isDebugEnabled()) {
-                                logBuilder = logBuilder.addKeyValue("responseBody", body);
-                            }
-                            logBuilder.log(LOGGER.isDebugEnabled()
-                                    ? "authentication operation failed"
-                                    : "authentication operation failed, increase log level to DEBUG for response body");
-                            throw new KmsException("%s failed with HTTP %d".formatted(operationType, response.statusCode()));
-                        }
-                    })
-                    .thenApply(HttpResponse::body)
-                    .thenApply(this::parseAuthResponse)
-                    .thenApply(authResponse -> {
-                        // Store refresh token for future use
-                        refreshToken.set(authResponse.refreshToken());
-
-                        // Calculate expiry time
-                        Instant now = Instant.now();
-                        Instant expiresAt = now.plusSeconds(authResponse.duration());
-
-                        LOGGER.atInfo()
-                                .addKeyValue("expiresAt", expiresAt)
-                                .log("{} succeeded", operationType);
-
-                        return new BearerToken(authResponse.jwt(), now, expiresAt);
-                    });
-        }
-        catch (IOException e) {
-            LOGGER.atWarn()
-                    .setCause(e)
-                    .log("failed to serialize authentication request");
-            return CompletableFuture.failedFuture(new KmsException("Failed to serialize authentication request", e));
-        }
-    }
-
-    private AuthResponse parseAuthResponse(byte[] bytes) {
-        try {
-            return OBJECT_MAPPER.readValue(bytes, AuthResponse.class);
-        }
-        catch (IOException e) {
-            String responseBody = new String(bytes, StandardCharsets.UTF_8);
-            var logBuilder = LOGGER.atWarn()
-                    .setCause(e);
-            if (LOGGER.isDebugEnabled()) {
-                logBuilder = logBuilder.addKeyValue("responseBody", responseBody);
-            }
-            logBuilder.log(LOGGER.isDebugEnabled()
-                    ? "failed to parse authentication response"
-                    : "failed to parse authentication response, increase log level to DEBUG for response body");
-            throw new UncheckedIOException("Failed to parse authentication response", e);
-        }
-    }
-
     @Override
-    public void close() {
-        client.close();
+    protected BearerToken createBearerToken(AuthResponse authResponse, String operationType) {
+        refreshToken.set(authResponse.refreshToken());
+
+        Instant now = Instant.now();
+        Instant expiresAt = now.plusSeconds(authResponse.duration());
+
+        LOGGER.atInfo()
+                .addKeyValue("expiresAt", expiresAt)
+                .log("{} succeeded", operationType);
+
+        return new BearerToken(authResponse.jwt(), now, expiresAt);
     }
 }
